refactor: modular entity handler
