+++ apps/publish-frt/base/src/builders/templates/mmoomm/playcanvas/handlers/EntityHandler/entityTypes/index.ts
@@ -1,0 +1,8 @@
+export { generatePlayerLogic } from './player';
+export { generateInteractiveLogic } from './interactive';
+export { generateVehicleLogic } from './vehicle';
+export { generateStaticLogic } from './static';
+export { generateShipLogic } from './ship';
+export { generateStationLogic } from './station';
+export { generateAsteroidLogic } from './asteroid';
+export { generateGateLogic } from './gate';
